git-ptx-patches: fix formating patches when patches from other tools are involved
If patches from another tool are involved, then series.append is created
with a "base" tag entry and an appropriate "--start-number". In this
case the "tags" array is not empty even if there is no "base" tag in the
actual series file and git-ptx-patches aborts.

So don't fill "tags" from series.append initially but set it later with
the new "base" tag if necessary.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>