rauc: fix PTXCONF_RAUC_CREATE handling
The existing --disable-create/--disable-create configure options were
not removed when the PTXCONF_RAUC_CREATE kconfig was introduced.

As a result of this, create could not be enabled for the target (since
it was forced to --disable-create afterwards) and the --enable-create
line was duplicated within the host-rauc recipe.

Fixes: c2ad273712cfb89c292dd5a72e8e1ee87a997ab7
("RAUC: leverage --disable-create option to shrink size for target RAUC")

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Message-Id: <20230630204927.2700230-1-ejo@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>