rauc: migrate to meson
Since different archive paths are used for the autotools and the meson
variant of rauc, we need to put them to distinct SRCDIR locations in
order to differentiate them when picking up. For this, add the -meson
suffix to RAUC_SOURCE.
This might be removed again when switching to the next release.

Also, RAUC_CONF_ENV modification does not seem to be needed since all
installed paths look valid without it.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Message-Id: <20230630210216.2724278-1-ejo@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>