ptxd_make_world_cargo_sync_package: add part of the git hash to the package source name
It not guaranteed that the package name and version are unique here. The
source URL may change, e.g. to a different git commit without changes to
the version.
As a result, the source package path remains the same, but the content
should change.

This happens for example with gstreamer-rs in gst-plugins-rs between
1.22.4 and 1.22.5.

So add part of the hash to the package file name to avoid this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>