kernel: add make target for the installed device trees
This makes it easy to depend on the kernel.targetinstall stage by
depending on one of the .dtb's that it installs into the image dir, e.g.
when building a custom recipe for a FIT image. As a nice side-effect,
you can simply call "ptxdist image <board.dtb>" to build the device
tree.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20230912165759.3492269-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>