nginx: version bump 1.18.0 -> 1.24.0
Modernize the systemd based on the upstream example from
https://www.nginx.com/resources/wiki/start/topics/examples/systemd/
With PrivateTmp=true, the tmpfiles.d part for /var/tmp/nginx no longer
works, so remove it and create the base directory in the service.
The other dirs with be created automatically.

configure options sorted according to the --help output.

Updated patches from buildroot.

Only the copyright years changed in the license file.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>