alsa-utils: fix build with(out) aplaymidi
If alsa-lib is built with ALSA_LIB_SEQ enabled, alsa-utils will always
build aplaymidi, and it will try to use a header alsa/ump_msg.h that is
only installed with ALSA_LIB_RAWMIDI enabled. Select ALSA_LIB_RAWMIDI
to install the missing header when aplaymidi is going to be installed
via ALSA_UTILS_MIDI, and set an override to stop it from including the
missing header otherwise.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Message-Id: <20231009124657.2869152-1-p.zabel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>