libxcrypt: new package
Also implement the needed logic to (optionally) replace
the libcrypt from the selected libc with libxcrypt.

libxcrypt is a modern library for one-way hashing of passwords.
It supports a wide variety of both modern and historical hashing
methods: yescrypt, gost-yescrypt, scrypt, bcrypt, sha512crypt,
sha256crypt, md5crypt, SunMD5, sha1crypt, NT, bsdicrypt, bigcrypt,
and descrypt. It provides the traditional Unix crypt and crypt_r
interfaces, as well as a set of extended interfaces pioneered by
Openwall Linux, crypt_rn, crypt_ra, crypt_gensalt, crypt_gensalt_rn,
and crypt_gensalt_ra.

libxcrypt is intended to be used by login(1), passwd(1), and other
similar programs; that is, to hash a small number of passwords
during an interactive authentication dialogue with a human.  It is
not suitable for use in bulk password-cracking applications, or in
any other situation where speed is more important than careful
handling of sensitive data.  However, it is intended to be fast and
lightweight enough for use in servers that must field thousands of
login attempts per minute.

Co-authored-by: Björn Esser <bes@pengutronix.de>
Signed-off-by: Björn Esser <bes@pengutronix.de>
Signed-off-by: Andreas Helmcke <ahelmcke@ela-soft.com>
Message-Id: <358b0e0c-b888-41f1-a21f-f9fa4d15b0fc@ela-soft.com>
[mol: only use it with glibc and add dependencies]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>