image-rauc: pick up hook script if it exists
Introduce a new kconfig option to package a hook script into the RAUC
bundle and reference it in the manifest. Otherwise leave the [hooks]
section in the manifest empty. The hook script is searched
config/images/rauc-hooks.sh in the usual search locations. RAUC will
complain if a hook script is referenced but no hook script is found in
the image (i.e. IMAGE_RAUC_HOOK_SCRIPT is empty).

As hooks that should be called by RAUC need to be mentioned in the
manifest too (in the relevant slot or the [bundle] section), the image
config will still need to be forked by BSPs in order to adapt the
manifest for that use case. However, the mechanism to find the hook
script is then already in place, and can be used as-is.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20231114220708.2816211-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>