linux-firmware: Version bump. 20231030 -> 20231111
A shorter release cycle. But otherwise, the usual stuff.
GPU, radio stuff. Nvidia firmware for Turing and later included.

A fix in the copy-firmware.sh script forces installation to be
dependent on the rdfind (redundant data find) binary.
I can't see that it's strictly needed nor do I think it's
a good idea to depend on "exotic" binaries for installation.
The script doesn't fail even if it doesn't find rdfind, so
just let it complain and go on with life. Like it's done
in the previous releases. I've voiced my opinion to a
linux-firmware maintainer.

* Add license for the old Intel ixp4xx series.

* Add patch to remove strict dependency on finding rdfind.
https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/diff/copy-firmware.sh?id=cf8315ded9b422d2e8b620ffe8bb661221639c8f

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20231115090400.2154434-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>