git-ptx-patches: Fix a race condition
As the loop over the output of find moves files around, it can happen
that find lists files more than once. In that case not only the "From "
line is removed, but also the next one (usually the "From:" line).

To process each patch in series.auto exactly once, iterate over that
file to remove the "From " line instead. Still be a bit conservative and
make sure that only the intended line is dropped.

Reported-by: Thorsten Scherer <t.scherer@eckelmann.de>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Tested-by: Thorsten Scherer <t.scherer@eckelmann.de>
Message-Id: <20231129093743.906378-2-u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>