igt-gpu-tools: avoid installing all files in targetinstall
The igt-gpu-tools installed all files by using install_tree during
targetinstall. That resulted in the installation in a lot of
platform-dependent tools and tests to be installed.

Avoid this by explicitly installing the tools and tests depending on the
package configuration.

As code_cov_parse_info is a Perl script, add Perl as a runtime
dependency if the tests are enabled.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20231204144934.426001-2-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>