host-dtc: Move devicetree compiler from platform to hosttools_noprompt.
The dtc has traditionally been a part of the platform side.
It was selected with a specific version to compile various kernel dts:es.
Nowdays, target programs has started to require the libfdt in runtime.
Ie, programs like QEMU. Beside the runtime, QEMU also have a
host-variant which need host tools.

This work preps for the QEMU upgrade by using the previous work
of target DTC to provide HOST_DTC as a normal ptxconf hosttool.

The end result is that HOST_DTC is reachable for HOST_QEMU.
This also removes the specific version for the platform side.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20240122183341.4113189-1-christian.melki@t2data.com>
Acked-by: Denis Osterland-Heim <denis.osterland@diehl.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>