optee-client: install PKCS#11 module libckteec
On next update, the imx platform will enable CFG_PKCS11_TA by default
and optee-test will fail to build, because we don't install the
respective PKCS#11 module.

PKCS#11 is important enough of a use case that it makes sense to just
ship it always by default, so let's do just that.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Message-Id: <20240126195301.982623-1-a.fatoum@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>