readline: fix pkg-config errors with NCURSES_WIDE_CHAR
readline.pc contains Requires.private=ncurses, even if readline is built
with ncursesw. When a user of readline uses pkg-config to ask for the
cflags, the lookup fails if ncursesw without compatibility with ncurses
is used.

Fix it by ensuring that ncurses compatibility is enabled if readline is
built with ncurses.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20240130105815.3091486-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>