wrapper: change -isystem<path> to -isystem <path> for icecc
icecc only handles '-isystem <path>' not '-system<path>' (without
space). As a result, the argument is not detected as local flag.

Meson generates '-system<path>' and in mesa 24.0.1, the sysroot-target
include is sorted before a local 'system' include path for gtest. That
causes the wrong header file to be used and building fails.

So add the space to work around this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>