ptxdist: drop sticky bits for the platform directory and sysroot-*
The sticky bit is inherited even if the mode is explicitly set when the
directory is created. As a result, the sticky bit is propagated through
the entire directory tree. That causes problems because it will change
the result of install_tree, causing conflicting directory permissions.

So drop the sticky bit for the platform directory and therefore for the
rest of the tree.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>