canfestival: port to Python 3
The gnosis library is extracted and moved around by the objdictgen
Makefile. Extract it early and do the same moving-around in the extract
stage so we can patch it in PTXdist.

Not all of the Python code was ported, only enough to make the build
work, which calls objdictgen.py to generate the C code for the examples.
The examples are fairly extensive, so this should work for most
user-supplied XML schema definitions. Of gnosis, only the XML pickle
modules and the introspection module was ported since those are the only
modules used by objdictgen. The test cases were mostly ignored, and some
of them that test Python-specific class internals also don't apply any
more since Python 3 refactored the whole type system. Also no care was
taken to stay compatible with Python 1 (duh!) or Python 2.

Upstream is apparently still dead, judging from the Mercurial repo (last
commit in 2019), the messages in the SourceForge mailing list archive
(last message in 2020, none by the authors), and the issue tracker (last
in 2020, none by the authors). gnosis is a whole different can of worms
which doesn't even have a publicly available repository or contact
information. So no attempt was made to send the changes upstream.

Remove a comment which referenced the old repository URL, which no
longer exists, and remove the recipe from staging.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20240312103109.3581087-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>