host-system-python3: make sure we use our own setuptools
The system python3 venv already contains setuptools, but it is an
undefined version. So remove it and install our own packag to ensure we
use a defined version.
This also avoids a race condition during parallel building: Installing
one of these packages while setuptools is installed may fail if Python
is executed when some of the files from setuptools are still missing in
the venv.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>