host-meson: ignore uic if it is not found
Qt builds and installs the uic only if Widgets are enabled, but meson
always checks for the uic when checking for the Qt tools. This breaks
the build of applications that use Qt, but don't use Widgets.

Add a patch to ignore the uic if it is not found, as PTXdist already
correctly handles the dependencies.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20240315095623.1888719-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>