ptxd_install_generic: drop set-group-ID for directories with explicit group
If the group is explicitly overwritten and not used from the directory
that is copied, then it does not make sense to keep the set-group-ID bit
meant for a different group.

This is relevant for install_alternative_tree. If the directory where a
BSP is cloned has the set-group-ID bit set, then the whole source tree
inherits it, including all directories in projectroot/. As a result
install_alternative_tree would copy it.
In such a use-case install_alternative_tree is typically called with
group=0 (or another explicitly specified group) so that can be used as
indicator when the set-group-ID bit should be dropped.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>