ptxd_kconfig_update: don't forget changes made by the migrate helper
Updating the stamp here is done so that updated to the .config timestamp
done by the migrate helper will not always cause the config to be saved.
However, if the migrate helper actually changes something in the file
then don't update the timestamp to ensure that the config is actually
updated with those changes.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>