migrate_ptx: keep the config stable when initmethod is not systemd
This is a bit tricky. In the initmethod systemd case, SYSTEMD and
SYSTEMD_HELPER are both set (new version) or only SYSTEMD is set.
In the non initmethod systemd case, SYSTEMD_HELPER is enabled or
disabled (new version) or SYSTEMD is enabled or disabled.

For the migration, SYSTEMD must be copied to SYSTEMD_HELPER but without
generating a duplicate SYSTEMD_HELPER symbol.

The current version just drops SYSTEMD_HELPER but that breaks the non
initmethod systemd case.

The only way to fix that is to evaluate the two lines together:
If both exist, then just print them (initmethod systemd new version).
If the second line is not SYSTEMD then print the first line and restart
with the second line (non initmethod systemd new version).
In all other cases, generate SYSTEMD_HELPER from SYSTEMD (old version).

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>