spirv-headers: version bump 1.3.275.0 -> 1.3.280.0
For license change, see upstream commit 1c9115b562ba ("List all licenses
in the root LICENSE file. (#410)"). The MIT license was added for
spir-v.xml and for the embedded jsoncpp.

The custom Khronos license governing most of the files is similar to the
MIT license, but with an added section about modifications making the
resulting files no longer reflect Khronos standards.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Message-Id: <20240326082221.2458163-2-p.zabel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>