rauc/image-rauc: move BUNDLE_VERSION to image-rauc
RAUC_BUNDLE_VERSION is only used by the image rule to generate the
bundle. Therefore, it belongs into image-rauc.

This fixes the issue that RAUC is rebuilt on every commit in a BSP, as a
commit changes PTXDIST_BSP_AUTOVERSION, which in turn changes the
configuration of RAUC. The rebuild is not necessary and only the RAUC
bundle should be rebuilt if the version has changed.

Rename the config, as the move to the image moves the item from the
ptxconfig to the platformconfig.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20240410081234.3902043-1-m.tretter@pengutronix.de>
[mol: fix indention]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>