vim: allow vim to be installed alongside busybox vi and xxd
Commit ad2eb34a07ed ("vim: prevent file name conflicts with busybox")
prevented vim being selected if busybox vi is selected. However, the vim
package does not currently install /usr/bin/vi, so the only conflict is
between vim's xxd and busybox's xxd.

Allow vim to be installed alongside busybox vi. Also add a configuration
option to allow installation of a /usr/bin/vi symbolic link to
/usr/bin/vim, but only allow it to be selected if busybox vi is not
selected.

Make the auto-selection of LIBC_M depend on VIM_VIM because VIM_XXD does
not appear to require it.  While we are at it, remove the obsolete
GCCLIBS_GCC_S and LIBC_DL selections.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Message-Id: <20240503180814.636648-1-abbotti@mev.co.uk>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>