ptxd_make_world_inject: Introduce new variable <PKG>_INJECT_OOT
Setting the new variable to 'YES' allows to inject into <PKG>_BUILD_DIR
instead of <PKG>_DIR as before.

When building out-of-tree some bootloaders like U-Boot expect injected
files in the build dir, not in the source dir.  Backwards compatibility
is ensured, the source dir is still the default, the new inject dest is
optional and in case set it overwrites the default.

For using this in u-boot package, on top of the things already done to
use the old inject mechanism in general, add this line to
'rules/u-boot.make' for example:

    U_BOOT_INJECT_OOT     := YES

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Acked-by: Michael Riesch <michael.riesch@wolfvision.net>
Message-Id: <20240506105051.323546-5-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>