gdb: Add option to configure lzma support
GDB's '--with-lzma=auto/yes/no' option defaults to 'auto' but it would
be better to configure it to 'yes' or 'no' explicitly.  GDB's lzma
support is required to read the .gnu_debuginfo sections in binaries (if
they have been built to include those sections), so call the
configuration option GDB_DEBUGINFO_SUPPORT.

The '--with-lzma' option was added in GDB 7.6, but that is quite old, so
assume it is supported.

While we are at it, change the test `ifneq ($(filter
1%,$(GDBSERVER_VERSION)),)` to use `$(GDB_VERSION)` instead, and add the
'--without-python' configure command line option.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Message-Id: <20240508101404.272507-1-abbotti@mev.co.uk>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>