bash: allow installation of /usr/bin/sh symlink to be chosen manually
It is possible for any of BusyBox, Dash, or Bash to install the
/usr/bin/sh symbolic link.  Only allow BASH_SH to be enabled if the
symbolic link is not being installed by either BusyBox or Dash (unless
overridden by ALLYES).  Default to enabled if the link is not being
installed by BusyBox or Dash.  Also allow it to be manually disabled.

Rationale: It is common for Bash to be installed as an interactive login
shell, but to use something lighter and faster for non-interactive use.
Shell scripts typically use /bin/sh as the interpreter selected by the
"shebang" line.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Message-Id: <20240513132247.25521-3-abbotti@mev.co.uk>
[mol: simplify 'default']
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>