kernel/perf: enable jevents
perf jevents are metrics that build on top of the raw perf events and
can provide much more insight at a glance than the raw numbers. Enable
them by default as the only dependency is host-python to generate the
C files to build into the perf target tool from the json description.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Message-Id: <20240531195144.1849501-1-l.stach@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>