ptxd_make_world_report_yaml_fragment: use correct name for the main package
For the fragments, the lower-case package name is generated. For the
first fragment (which is usually the main package), this should match
the actual lower-case package name.
Unfortunately there are some old packages that use '_' in the lower case
name, so the resulting name does not match.
This confuses the SBOM generator.

So special case the main packge and use pkg_label.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>