templates: kernel: use ptx/image-install and fix clean stage
The clean stage did not clean up the installed kernel images from the
image directory, but if we use ptx/image-install to install them, this
can all be done by PTXdist's default clean stage. That macro now also
prints the installed images, so we don't need to echo them.

While at it, fix a misplaced space in the "foreach dtb" statement.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20240808093643.1634147-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>