mesalib: read libgallium version from VERSION file
Use the version from the VERSION file to determine the libgallium file
name. Fall back to using MESALIB_VERSION if the sources are not present.

When building from git, the VERSION file may contain a version in the
form 24.3.0-devel, which is likely different from MESALIB_VERSION. The
build system determines the libgallium file name from the VERSION file,
so we should do the same.

This fixes targetinstall for git builds, e.g. when using local-src. For
release tarballs, the VERSION file content and MESALIB_VERSION are
identical.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Message-Id: <20240815115615.2745801-3-p.zabel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>