nginx: Allow ngx_http_charset_module to be built
Add a configuration option to build the http charset module. It allows a
charset to be specified in the HTTP "Content-Type" header, for example
for setting the charset to "utf-8".

It can also do charset conversions with some limitations, but the
nginx.make file has not been modified to install any of the supplied
conversion tables.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Message-Id: <20240830132722.65713-1-abbotti@mev.co.uk>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>