libcurl: Version bump. 8.9.1 -> 8.10.0
Bunch of changes as usual.
https://curl.se/changes.html#8_10_0

The most interesting part is probably the removal of
selection of random file (selected as urandom in ptxdist).
The said scheme in curl is supposed to cover the situation.
Let's hope it works out fine for most situations and things
doesn't hang on low entropy scenarios.

* Pin down a few options.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20240912192840.1752041-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>