llvm: fix and cleanup cross-compiling
The target llvm must provide a llvm-config that reports the correct
configuration, such as the build components etc.

That was provided by the host-llvm package but that conflicts with the
requirements of that package itself. And disabling unused target
(1532271f1003d236520b9b6578f3d8fb85dca9db "host-llvm: drop all targets")
broke that.

Instead, introduce a new cross-llvm package that only builds llvm-tblgen
and llvm-config with basically the same configuration as the target llvm
package. And then use the llvm-config from this package.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>