ptxd_make_nested_ptxdist: allow different PTXdist versions for nested BSPs
If the nested PTXdist call uses a different workspace, then it is most
likely an entirely different BSP. So it can have a different PTXdist
version.
Support that and let --auto-version handle the version switch.

To make that possible, PTXDIST_PLATFORMDIR must be cleared first.
The problem is this:

We keep PTXDIST_PLATFORMDIR set to detect if the nested ptxdist call
builds the same platform and abort with an error.

This worked before with --autoversion because normally
PTXDIST_PLATFORMDIR is not exported before the new PTXdist version is
executed, so that value is not propagated.

However, if it is already exported in the nested case then the value is
updated and the PTXdist called by --autoversion will incorrectly detect
a nested build for the same platform.

To avoid this, make sure that PTXDIST_PLATFORMDIR is not exported,
1. when the workspace is different and PTXDIST_PLATFORMDIR will never
   match anyways
2. when setting the new PTXDIST_PLATFORMDIR.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>