ptxdist: remove superfluous deletion of logs in distclean action
There is always a platform build directory, the logfile is not created
in BSP root ($PTXDIST_WORKSPACE/logfile) anymore since commit
83bf2580f8 ("[ptxdist] always use a platform dir", 2010-11-15).

Furthermore, $PTX_LOGFILE is already deleted in clean() invoked at the
beginning of the "distclean" action, so the whole rm command is
unnecessary.

Also remove "logfile" from ignore rules, it is only created in the
platform build directory of the BSP, not in the PTXdist tree.

Signed-off-by: Markus Heidelberg <m.heidelberg@cab.de>
Message-Id: <20250103150647.136289-2-m.heidelberg@cab.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>