libtracefs: fix license and add license files
While the source includes a GPL-2.0 license text, this license is only
used for the build scripts, and does not make it into the binary. The
Bison-generated parser code is licensed under GPL-3.0-or-later with a
license exception allowing it (in this case) to be released under
a different license. The resulting binary is under LGPL-2.1-or-later (as
mentioned in the original mail thread it uses the old SPDX identifiers).

Link: https://lore.ptxdist.org/ptxdist/Z4DbKik6UoE956Sq@pengutronix.de
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Message-Id: <20250117104113.176520-1-rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>