gptfdisk: add new package
GPT fdisk is a set of tools to work with GUID partition tables: gdisk,
cgdisk, sgdisk and fixparts

Of particular interest to embedded devices is sgdisk, the noninteractive
tool, which, amongst other functionality, allows relocating the backup
(alternative) GPT header to the end of the disk. This may be needed
after flashing a GPT disk image created for a smaller disk size e.g. to
account for fluctuation in disk size amongst hardware revisions.

The curses based cgdisk makes use of the wide char ncurses variant. As
ptxdist currently, unlike some linux distributions, does not set up
ncursesw/ncurses.h as a symlink to ncurses.h, gptfdisk is patched to
include ncurses from the standard path.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Message-Id: <20250131-gptfdisk-v2-1-c6de969b69bb@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>