ptxd_make_world_cargo_sync: handle the same package/version with multiple revisions
The same package and version but with a different revision can be used
as dependency. Detect and handle this so that both are available.
Without this, the same variable names are used and the second instance
overwrited the first one.

Take that into account as well when creating replace-with entries and
deduplicate without the revision.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>