kernel/perf: enable zstd support
The Linux kernel Commit 44b44ffd5dce ("perf build: Minor improvement for
linking libzstd") adds a test if libzstd is installed and uses this as
heuristic if libelf is build with zstd support. Thus, if zstd is
installed, perf cannot be built without zstd support anymore.

Always build perf with zstd support and make sure that libelf is built
with zstd, too.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20250317113146.1862763-2-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>