glibc: install libmvec only if supported by architecture
If libmvec is available depends on the architecture and the glibc
version. Thus, it is not available on all architectures.

This is a problem on a BSP that contains platforms with different
architectures, if some of the architectures have libmvec and others
don't.

Respect PTXCONF_GLIBC_MVEC only on architectures that actually have
libmvec and ignore the setting otherwise.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20250411142222.2721457-1-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>