ptxd_kconfig_validate_config: fix validation
Due to a typo, the config was not actually validated. So changes in the
base layer may not be detected.

There are more things broken that where hidden by the typo and are fixed
here:

Remove an incorect local variable to ensure that 'next' is actually set.

"${@}" are now the configs not the layers, so for the _REF_CONFIG
fallback, the serach path must be modified accordingly.

Fixes: 47d24525a49d ("ptxd_kconfig_find_config: refactor to define the list of possible configs first")
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>