readline: Install libhistory.
libhistory is part of readline and provided as a separate .so.
It's fairly small (about 1/10:th of readline and around 40k on my machine).
I don't see a big point in ignoring to install it, esp. since there is no
builtin option to turn off the header installs and similar.

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20250513192818.3075825-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>