libconfig: version bump 1.7.2 -> 1.8
Switch to the git tarball from github. The tarball for the old version
was removed from the old URL, so we can expect the same thing to happen,
wenn the next version is released.
Avoid that by using the git tarball und run autogen.sh at build-time.

Copyright year update in the license header.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>