networkmanager: switch from gnome mirror to freedesktop.org
The latest network manager releases are not available on the gnome
mirror file servers. At the time of testing, the latest release in the
FDO gitlab is 1.53-dev, while the gnome mirror only has 1.51.

Switch download to the freedesktop.org gitlab. It does not provide xz
archives, so we use the bz2 files instead.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>
Message-Id: <20250521090234.2324640-1-r.schwebel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>