ptxd_kconfig_validate_config: don't break if config exists on all layers
If the config exists on the lowest layer, then
ptxd_kconfig_validate_config_next() is called with out arguments, so
next is not updated at all. and ptxd_kconfig_validate_config() expects
another layer and fails because the diff is missing.
Avoid this by clearing next in this case.

This happens with 'ptxdist localsetup' where the config in the last
layer is ptxdistrc.default.

Fixes: 30afb3fa3753 ("ptxd_kconfig_validate_config: fix validation")
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>