openssh: Non-change version bump. 10.0p1 -> 10.0p2
Apparently the p1 release name was a mistake and name is changed to p2.
Although the same tarball exists as p1 (no hash change), change
the tarball name to p2 to remove any chance of confusion.
I didn't chatch this mistake at the time of the last version bump,
so lets fix this now.

https://seclists.org/oss-sec/2025/q2/35

Signed-off-by: Christian Melki <christian.melki@t2data.com>
Message-Id: <20250525170439.2084354-1-christian.melki@t2data.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>