ptxdist: improve platformconfig selection
If ptxdist finds only one platformconfig, it will use that by default.
With multiple layers, this is a bit more complex:
ptxdist walks through all layers:
 - no platformconfig found: go to the next layer
 - only one platformconfig found: use it
 - more than one platformconfig found: only use selected_platformconfig
   if available

This does not work correctly in all cases: If the top layer has multiple
platformconfigs, it is skipped. If a lower layer has only one
platformconfig, this config is incorrectly used.

So if multiple platformconfigs are found, restrict the search to
selected_platformconfig only for lower layers. This way no default
platform is found.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>