wpewebkit: version bump 2.42.5 -> 2.49.3
Building wpewebkit with ENABLE_VIDEO=OFF is broken. There are several
locations that unconditionally use classes that are hidden behind `if
ENABLE(VIDEO)` and the build exposes various missing definitions.  As a
workaround prevent the build with ENABLE_VIDEO=OFF.

Disable ENABLE_WPE_PLATFORM_DRM, ENABLE_WPE_PLATFORM_HEADLESS, and
ENABLE_WPE_PLATFORM_WAYLAND since they only have an effect if
ENABLE_WPE_PLATFORM is ON, which is disabled as an experimental feature.
The WPE Platform API is an alternative to libwpe and the WPE backends.

Add libdrm as dependency and enable USE_LIBDRM, since USE_GBM has been
split in USE_GBM and USE_LIBDRM to better handle code that requires DRM
but not GBM.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Message-Id: <20250711081140.4128890-4-m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>