ptxd_lib_dgen: fix report stage with devpkgs
The report stage depends on the targetinstall stage for target packages
If a devpkg is found, then it is extracted and used instead of building
from the source. In this case, the extract stage is never executed and
the report stage fails because it cannot find the license files.

To fix this, let the report stage explicitly depend on the extract stage
if a devpkg is found.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>