libubootenv: Add option for installing tools
These tools can also be provided by the u-boot-tools package.

This variant promises to be board independent, and comes with a library,
and has more features.

Can be used as drop-in replacement for fw_printenv and fw_setenv.
Helpful for writing env to mmcblockXbootX which is ro by default, and
u-boot-tools does not handle that.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Message-Id: <20250919100643.199174-3-ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>