ldd: switch to version provided by glibc
Until now, the ldd version installed on the target has been a script
copied from projectroot. Besides not having been updated since v2.7
(we are at v2.40+ now), this approach does not account for the various
architectures PTXdist can build BSPs for.

We already have access to a recent version of ldd, as it comes with
glibc from the active toolchain. We can hence drop the local ldd script
and directly install the tool from there.

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Message-Id: <20251008051341.937149-1-h.assmann@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>