opkg: version bump 0.7.0 -> 0.9.0
Due to access problems, the current releases of opkg are not currently
uploaded to the release folder. Therefore get it from the git repo. See:
https://lists.yoctoproject.org/g/opkg/topic/release_tarball_for_0_9_0/115374634

opkg switched in version 0.9.0 to CMake as a build system.

Use the libsolv library, as the internal SAT solver is marked as
deprecated and will probably be removed in a future version.

https://git.yoctoproject.org/opkg/tag/?h=v0.8.0
https://git.yoctoproject.org/opkg/tag/?h=v0.9.0

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251013141237.2372882-2-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>